src/requirements.txt
dash==2.17.1
dash-extensions==0.1.13
dash-local-react-components==1.3.0
litellm==1.52.14
dash-mantine-components==0.14.5
dataclass-wizard==0.22.2
python-dotenv==1.0.0
